+++ modules/prelude-helm-everywhere.el
@@ -49,17 +49,11 @@
 
 (define-key isearch-mode-map (kbd "C-o") 'helm-occur-from-isearch)
 
-;; shell history.
-(define-key shell-mode-map (kbd "C-c C-l") 'helm-comint-input-ring)
-
 ;; use helm to list eshell history
 (add-hook 'eshell-mode-hook
           #'(lambda ()
               (substitute-key-definition 'eshell-list-history 'helm-eshell-history eshell-mode-map)))
 
-(substitute-key-definition 'find-tag 'helm-etags-select global-map)
-
-(helm-descbinds-mode)
 (helm-mode 1)
 
 (when prelude-projectile